[dinosaur] remove StateWithTime
Instead of two different versions of primitive equations state (State/StateWithTime) and equation classes (PrimitiveEquations/PrimitiveEquationsWithState), make `sim_time` an optional attribute of `State`, and use unified equation classes.

This will allow users to avoid including sim_time in cases where they don't use it, and therefore avoid misleading uses of the name `StateWithTime` with invalid sim_time data. It's also just bit simpler overall.

In the long term, we might consider dropping the propagation of sim_time as an ODE term entirely (it is very suspectible to round-off errors) but this would break backwards compatibility with current NeuralGCM models and thus is probably not the worth the trouble.

PiperOrigin-RevId: 711880923